pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.2",
    "groq>=0.36.0",
    "langchain-community>=0.3.31",
    "livekit-agents[google,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "pydantic>=2.12.4",
    "python-dotenv",
    "requests>=2.32.5",
    "typer>=0.20.0",
    "typer-slim>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.workspace]
members = [
    "browser-use-node",
]
